# Don't reparse large integers