This handles overflow checking without reparsing. It should significantly reduce the number of instructions in the program, as well as instructions executed when things go bad. It also eliminates a never_inline function, which is always a happy thing.

## Algorithm

It first separates negative and positive overflow checking (which, whether a number is negative is actually known at compile time due to detection of the initial `-`).

Negative numbers are simple:
* If it's more than 18 digits, it has overflowed.
* If it's greater than INT64_MAX+1, it has overflowed (won't fit). uint64_t can fit all 18 digit numbers without overflow, so this is safe.

Positive numbers are harder:
* If it's more than 19 digits, it has overflowed.
* If it's *equal* to 19 digits, and the first digit is not '1', it has overflowed. (UINT64_MAX is 18,446,744,073,709,551,615.)
* If it's 19 digits, and the first digit is 1, and it's less than 10,000,000,000,000,000,000 (the smallest 19 digit number starting with 1), it has overflowed. (We actually check if it's less than INT64_MAX--which is a little shy of that minimum--for convenience.)

I spent a good deal of time making it as branchless as possible. I also moved float parsing out of the main body of the function into other really_inline functions, since it was getting a little big. No performance impact there.

### Overflow Behavior

That last rule is the only one that relies on any overflow behavior: specifically, it assumes that any number between 1,844,674,407,370,955,161 and 1,999,999,999,999,999,999 will produce a *smaller* number than 10,000,000,000,000,000,000. This seems to hold true on the machines we run on.

## Performance

I would have expected this to improve twitter, and it doesn't: twitter has a lot of 18-digit unsigned integers in it, which should be able to bypass the entire parse_large_integer in a way it didn't before.

This increases performance reasonably in a few files without making anything slower, on skylake:

| File                                | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                                 |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| mesh                                | 269.9 | 255.2 |    5% | 841.5 | 848.7 |    0% |   9917 |   9725 |    1% |      6 |      0 |    0% | 137176 | 127706 |    6% |
| numbers                             | 228.9 | 219.3 |    4% | 704.1 | 699.8 |    0% |   2124 |   2143 |    0% |      0 |      0 |    0% |  18659 |  17584 |    5% |
| marine_ik                           | 262.9 | 252.0 |    4% | 792.2 | 796.3 |    0% |  48748 |  49213 |    0% | 268251 | 277655 |   -3% | 555301 | 556907 |    0% |
| mesh.pretty                         | 166.1 | 160.3 |    3% | 512.1 | 515.4 |    0% |  13456 |  13285 |    1% |    348 |    101 |   70% | 181418 | 180880 |    0% |
| update-center                       | 113.9 | 111.8 |    1% | 338.7 | 339.5 |    0% |   6097 |   5552 |    8% |      0 |      0 |    0% |  79460 |  79626 |    0% |
| apache_builds                       |  88.2 |  87.8 |    0% | 307.8 | 309.2 |    0% |    110 |    108 |    1% |      0 |      0 |    0% |  13959 |  14104 |   -1% |
| google_maps_api_response            | 104.9 | 104.4 |    0% | 356.7 | 357.7 |    0% |     63 |     63 |    0% |      0 |      0 |    0% |    365 |    241 |   33% |
| google_maps_api_compact_response    | 196.6 | 195.8 |    0% | 674.2 | 676.4 |    0% |     64 |     63 |    1% |      2 |      0 |    0% |    157 |    154 |    1% |
| twitter                             |  91.8 |  91.4 |    0% | 291.3 | 292.2 |    0% |   3425 |   3334 |    2% |      0 |      0 |    0% |  78586 |  78550 |    0% |
| tree-pretty                         |  95.7 |  95.4 |    0% | 332.1 | 334.1 |    0% |     51 |     60 |  -17% |      0 |      0 |    0% |   1580 |    340 |   78% |
| canada                              | 275.9 | 275.1 |    0% | 894.5 | 889.8 |    0% |  18622 |  19237 |   -3% |  46520 |  41402 |   11% | 324387 | 323159 |    0% |
| gsoc-2018                           |  71.8 |  71.6 |    0% | 170.5 | 170.8 |    0% |  30107 |  30871 |   -2% | 102688 | 103640 |    0% | 352070 | 350784 |    0% |
| repeat                              | 105.6 | 105.6 |    0% | 354.3 | 356.1 |    0% |     14 |     13 |    7% |      0 |      0 |    0% |     39 |      4 |   89% |
| github_events                       |  80.2 |  80.3 |    0% | 265.9 | 267.6 |    0% |    112 |    100 |   10% |      0 |      0 |    0% |   3490 |   3279 |    6% |
| twitterescaped                      | 191.7 | 192.0 |    0% | 554.0 | 555.1 |    0% |   4952 |   4927 |    0% |      0 |      0 |    0% |  74974 |  75143 |    0% |
| instruments                         |  95.2 |  95.4 |    0% | 340.0 | 344.2 |   -1% |    499 |    419 |   16% |      0 |      0 |    0% |  29053 |  29698 |   -2% |
| twitter_api_compact_response        | 115.0 | 115.4 |    0% | 371.0 | 372.6 |    0% |     16 |     20 |  -25% |      0 |      0 |    0% |     78 |     34 |   56% |
| twitter_api_response                |  92.7 |  93.0 |    0% | 302.2 | 303.2 |    0% |     20 |     26 |  -30% |      0 |      0 |    0% |     95 |    106 |  -11% |
| random                              | 138.1 | 138.8 |    0% | 480.4 | 482.3 |    0% |   1976 |   1965 |    0% |      0 |      0 |    0% |  86683 |  86511 |    0% |
| twitter_timeline                    | 130.1 | 130.9 |    0% | 396.4 | 398.0 |    0% |    133 |    141 |   -6% |      0 |      0 |    0% |   1777 |   1325 |   25% |
| citm_catalog                        |  79.4 |  80.2 |    0% | 293.5 | 294.7 |    0% |   1754 |   1745 |    0% |     18 |   1517 | -8327% | 182500 | 181913 |    0% |