@jkeiser 2020-06-25
I do wonder if something here makes gcc10 specifically unhappy. I tested gcc9 on skylake.