@jkeiser 2020-06-26
On ampere, there doesn't appear to be a difference between master and the PR:

| File                                | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                                 |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| twitter_api_compact_response        | 275.0 | 273.5 |    0% | 204.7 | 204.7 |    0% |     60 |     53 |   11% |    122 |    114 |    6% |   3245 |   3237 |    0% |
| google_maps_api_compact_response    | 282.4 | 281.3 |    0% | 246.0 | 246.0 |    0% |    129 |    120 |    6% |    191 |    186 |    2% |   4950 |   4947 |    0% |
| twitter_api_response                | 265.6 | 265.4 |    0% | 189.8 | 189.8 |    0% |     70 |     70 |    0% |    240 |    231 |    3% |   4323 |   4339 |    0% |
| random                              | 374.1 | 374.1 |    0% | 269.1 | 269.1 |    0% |   3782 |   3783 |    0% |   8023 |   8020 |    0% | 189980 | 189959 |    0% |
| citm_catalog                        | 250.9 | 250.9 |    0% | 179.8 | 179.8 |    0% |    450 |    445 |    1% |  27032 |  27029 |    0% | 434793 | 434772 |    0% |
| twitter                             | 292.7 | 292.7 |    0% | 199.5 | 199.5 |    0% |   3296 |   3282 |    0% |   9913 |   9909 |    0% | 166865 | 166844 |    0% |
| mesh.pretty                         | 251.4 | 251.4 |    0% | 188.0 | 188.0 |    0% |     40 |     43 |   -7% |  24686 |  24684 |    0% | 430715 | 430712 |    0% |
| mesh                                | 266.8 | 266.9 |    0% | 228.4 | 228.4 |    0% |   1022 |   1028 |    0% |  11345 |  11344 |    0% | 254423 | 254412 |    0% |
| gsoc-2018                           | 224.8 | 224.9 |    0% | 145.0 | 145.0 |    0% |   9661 |   9765 |   -1% |  52037 |  52037 |    0% | 574138 | 573944 |    0% |
| numbers                             | 258.9 | 259.0 |    0% | 195.6 | 195.6 |    0% |    620 |    581 |    6% |   2387 |   2386 |    0% |  44093 |  44089 |    0% |
| repeat                              | 392.0 | 392.1 |    0% | 243.4 | 243.4 |    0% |     36 |     36 |    0% |    134 |    125 |    6% |   3381 |   3374 |    0% |
| instruments                         | 257.8 | 257.9 |    0% | 194.3 | 194.3 |    0% |    766 |    761 |    0% |   3482 |   3480 |    0% |  63888 |  63889 |    0% |
| google_maps_api_response            | 257.0 | 257.2 |    0% | 187.4 | 187.4 |    0% |     70 |     69 |    1% |    434 |    433 |    0% |   7268 |   7265 |    0% |
| tree-pretty                         | 258.3 | 258.5 |    0% | 191.3 | 191.3 |    0% |    118 |    121 |   -2% |    572 |    577 |    0% |   9904 |   9910 |    0% |
| canada                              | 261.0 | 261.3 |    0% | 211.2 | 211.2 |    0% |  11008 |  11000 |    0% |  35215 |  35214 |    0% | 748085 | 748079 |    0% |
| apache_builds                       | 257.3 | 257.6 |    0% | 189.8 | 189.8 |    0% |    443 |    421 |    4% |   2031 |   2029 |    0% |  35542 |  35508 |    0% |
| marine_ik                           | 268.5 | 268.9 |    0% | 237.2 | 237.2 |    0% |  12821 |  12918 |    0% |  46656 |  46655 |    0% | 1145104 | 1144095 |    0% |
| update-center                       | 259.6 | 260.0 |    0% | 192.9 | 192.9 |    0% |   4149 |   4162 |    0% |   8384 |   8378 |    0% | 153798 | 153844 |    0% |
| twitterescaped                      | 250.7 | 251.1 |    0% | 181.3 | 181.3 |    0% |   4342 |   4421 |   -1% |   8832 |   8824 |    0% | 145975 | 146090 |    0% |
| github_events                       | 251.6 | 252.1 |    0% | 178.2 | 178.2 |    0% |    242 |    259 |   -7% |   1061 |   1059 |    0% |  16465 |  16467 |    0% |
| twitter_timeline                    | 264.1 | 264.7 |    0% | 199.7 | 199.7 |    0% |    388 |    408 |   -5% |    698 |    693 |    0% |  13010 |  13032 |    0% |