@jkeiser 2020-06-26
@lemire hold off on looking; I just discovered that I've been running my perf comparison script against stage 1 only.