@jkeiser 2020-06-27
The failing performance test is westmere; twitter loses a little on westmere while a few others gain:

| File                                | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                                 |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| instruments                         | 136.6 | 134.7 |    1% | 503.9 | 501.1 |    0% |    459 |    396 |   13% |      0 |      0 |    0% |  31975 |  32357 |   -1% |
| update-center                       | 160.5 | 158.5 |    1% | 534.6 | 528.3 |    1% |   5972 |   5502 |    7% |      0 |      0 |    0% |  85961 |  85930 |    0% |
| gsoc-2018                           | 106.3 | 105.6 |    0% | 313.1 | 311.9 |    0% |  30390 |  30599 |    0% | 102863 |  98219 |    4% | 370217 | 370471 |    0% |
| tree-pretty                         | 137.4 | 136.6 |    0% | 499.5 | 495.1 |    0% |     66 |     58 |   12% |      0 |      0 |    0% |   1107 |    388 |   64% |
| numbers                             | 253.5 | 252.0 |    0% | 822.3 | 813.8 |    1% |   2211 |   2048 |    7% |      0 |      0 |    0% |  19646 |  19515 |    0% |
| google_maps_api_compact_response    | 261.4 | 260.7 |    0% | 924.0 | 917.2 |    0% |     66 |     65 |    1% |      0 |      0 |    0% |     94 |     75 |   20% |
| twitterescaped                      | 253.7 | 253.5 |    0% | 780.8 | 775.6 |    0% |   4873 |   4648 |    4% |      0 |      0 |    0% |  81500 |  81473 |    0% |
| random                              | 206.8 | 206.6 |    0% | 759.8 | 753.2 |    0% |   1920 |   2201 |  -14% |      0 |      4 |    0% |  93407 |  93514 |    0% |
| google_maps_api_response            | 145.8 | 145.7 |    0% | 523.8 | 520.7 |    0% |     69 |     69 |    0% |      0 |      0 |    0% |    287 |    504 |  -75% |
| twitter_timeline                    | 184.7 | 184.7 |    0% | 597.3 | 590.5 |    1% |    160 |    155 |    3% |      0 |      0 |    0% |   1433 |   1312 |    8% |
| github_events                       | 120.9 | 121.0 |    0% | 428.8 | 425.3 |    0% |    129 |    120 |    6% |      0 |      0 |    0% |   3127 |   3225 |   -3% |
| canada                              | 308.2 | 309.2 |    0% | 1007.0 | 1002.2 |    0% |  18643 |  18846 |   -1% |  43355 |  49822 |  -14% | 347631 | 347396 |    0% |
| repeat                              | 158.7 | 159.3 |    0% | 579.3 | 578.2 |    0% |     15 |     12 |   20% |      0 |      0 |    0% |      4 |     50 | -1150% |
| apache_builds                       | 130.3 | 130.9 |    0% | 480.8 | 477.7 |    0% |    119 |    124 |   -4% |      0 |      0 |    0% |  14738 |  15137 |   -2% |
| twitter_api_compact_response        | 158.7 | 160.3 |   -1% | 562.0 | 556.1 |    1% |     13 |     19 |  -46% |      0 |      0 |    0% |    118 |     79 |   33% |
| citm_catalog                        | 113.0 | 114.4 |   -1% | 425.4 | 423.8 |    0% |   1760 |   1774 |    0% |   2291 |    529 |   76% | 198763 | 198682 |    0% |
| twitter_api_response                | 130.8 | 132.7 |   -1% | 468.9 | 464.1 |    1% |     19 |     30 |  -57% |      0 |      0 |    0% |    130 |    139 |   -6% |
| twitter                             | 135.2 | 137.2 |   -1% | 476.3 | 471.7 |    0% |   3160 |   3641 |  -15% |      0 |      0 |    0% |  85448 |  84908 |    0% |
| mesh.pretty                         | 189.0 | 192.3 |   -1% | 621.7 | 624.5 |    0% |  13380 |  13149 |    1% |    113 |      4 |   96% | 197809 | 197683 |    0% |
| marine_ik                           | 292.1 | 297.7 |   -1% | 933.0 | 933.7 |    0% |  47668 |  48527 |   -1% | 281106 | 270493 |    3% | 574757 | 574245 |    0% |
| mesh                                | 288.0 | 296.4 |   -2% | 961.3 | 966.4 |    0% |   9931 |   9807 |    1% |      0 |      0 |    0% | 138482 | 138245 |    0% |